Removed images, added more mobile functionality
In this update I added in filters for the various pages to allow for ease of search. On top of that, I changed the way the rendering is done, starting to rid of the ugly data grid for better looking cards. I then added mobile support and began to rework some backend database calls.